pyproject.toml
[project]
name = "local-agent"
version = "0.1.0"
description = "本地 Agent 框架，基于 gemma-4-31B-it，支持多模态输入和工具调用"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.40",
    "accelerate>=0.27",
    "chromadb>=0.4",
    "pymupdf>=1.24",
    "Pillow>=10.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "sentence-transformers>=2.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
agent = "agent.cli.interface:main"
